extract git root function to misc/path.rs
