+++ rust/cli/src/misc/mod.rs
@@ -1,0 +1,1 @@
+pub mod path;
